Fixed the unauthorized_users_may_not_delete_threads() test.

+++ app/Http/Controllers/ThreadsController.php
@@ -130,22 +130,17 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-        try {
-
-            $thread->delete();
-
-            if(request()->wantsJson()) {
-                return response([], 204);
-            }
-
-            return redirect('/threads');
-
-        } catch (\Exception $exception) {
-
-            \Log::error($exception->getMessage());
-
-            return response([], $exception->getCode());
+        if($thread->user_id != auth()->id()) {
+            abort(403, 'Permission Denied');
         }
+
+        $thread->delete();
+
+        if(request()->wantsJson()) {
+            return response([], 204);
+        }
+
+        return redirect('/threads');
     }
 
     /**
